created functional prototype for interactable signs
