+++ Assets/Scripts/Interactable/Sign.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0b618665bd08cee488c7356db9da5113
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
