--- Assets/Scripts/NavigationMap.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2d43cc2232a70004ca05f4e11fc9e50f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
