--- Assets/Scripts/NavigationMap/NavigationMap.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fb4885aab6a9b9f41b68de439566392c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
